okie

+++ plugins/delete.py
@@ -9,7 +9,8 @@ async def check_up(bot):
     all_data = await get_all_dlt_data(_time)
     for data in all_data:
         try:
-           await bot.delete_messages(chat_id=data["chat_id"],
+          if data['_time'] < _time:
+            await bot.delete_messages(chat_id=data["chat_id"],
                                      message_ids=data["message_id"])           
         except Exception as e:
            err=data
